seperate functions and rename typo filename
